added tests for Qlist, refactored file structure
