+++ Questions/QuestionsApp.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Questions from './component/Questions.jsx';
+
+ReactDOM.render(<Questions />, document.getElementById('app'));
