+++ client/src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Product from '../../Product/component/Product.jsx';
-import Questions from '../../Questions/component/Questions.jsx';
+import QuestionsApp from '../../Questions/QuestionsApp.jsx';
 import Related from '../../Related/component/RelatedItems.jsx';
 import Reviews from '../../Reviews/component/Reviews.jsx'
